Refactor led_interrupt.cpp
Due to author's ungodly mess in code I segregated functions into hpp and
cpp files. Names of variables were unclear.
